+++ scripts/_locations.ts
@@ -1,4 +1,4 @@
 export const DIR_ROOT = `${__dirname}/..`
-export const DIR_BOILERPLATES = `${DIR_ROOT}/create-vite-plugin-ssr`
+export const DIR_BOILERPLATES = `${DIR_ROOT}/boilerplates`
 export const DIR_EXAMPLES = `${DIR_ROOT}/examples`
 export const DIR_SRC = `${DIR_ROOT}/src`